+++ examples/service-sameness/cluster/variables.tf
@@ -1,0 +1,7 @@
+# Copyright (c) HashiCorp, Inc.
+# SPDX-License-Identifier: MPL-2.0
+
+variable "name" {
+  description = "Name of the ECS cluster."
+  type        = string
+}